refactor: dikstra.ts
